### 2023-11-07 #5882 @jedbrown
@dc-luo :wave: Hi, and sorry for the slow start on my part. For the paper build error, I think you just need to replace `"1 3"` with `"1, 3"`.

I've installed following the [developer instructions](https://hippylibsoupy.readthedocs.io/en/latest/install.html#installation-for-developers) and can run sequential poisson and hyperelasticity, albeit with many syntax warnings (conda uses python-3.12).

```
/home/jed/joss/hippylib/hippylib/modeling/expression.py:18: SyntaxWarning: invalid escape sequence '\s'
  '''
/home/jed/joss/hippylib/hippylib/algorithms/linalg.py:176: SyntaxWarning: invalid escape sequence '\s'
  """
/home/jed/joss/hippylib/hippylib/algorithms/lowRankOperator.py:58: SyntaxWarning: invalid escape sequence '\m'
  """
/home/jed/joss/hippylib/hippylib/algorithms/lowRankOperator.py:81: SyntaxWarning: invalid escape sequence '\m'
  """
/home/jed/joss/hippylib/hippylib/algorithms/lowRankOperator.py:106: SyntaxWarning: invalid escape sequence '\m'
  """
/home/jed/joss/hippylib/hippylib/algorithms/cgsampler.py:21: SyntaxWarning: invalid escape sequence '\m'
  """
/home/jed/joss/hippylib/hippylib/algorithms/cgsampler.py:53: SyntaxWarning: invalid escape sequence '\s'
  """
/home/jed/joss/hippylib/hippylib/modeling/misfit.py:60: SyntaxWarning: invalid escape sequence '\d'
  """
/home/jed/joss/hippylib/hippylib/modeling/misfit.py:241: SyntaxWarning: invalid escape sequence '\s'
  """
/home/jed/joss/hippylib/hippylib/modeling/misfit.py:246: SyntaxWarning: invalid escape sequence '\O'
  """
/home/jed/joss/hippylib/hippylib/forward_uq/qoi.py:50: SyntaxWarning: invalid escape sequence '\d'
  """
/home/jed/joss/hippylib/hippylib/forward_uq/variationalQoi.py:65: SyntaxWarning: invalid escape sequence '\d'
  """Apply the second variation \delta_ij (i,j = STATE,PARAMETER) of the cost in direction dir."""
/home/jed/joss/soupy/examples/poisson/../../soupy/modeling/controlCostHessian.py:57: SyntaxWarning: invalid escape sequence '\h'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/modeling/controlModel.py:154: SyntaxWarning: invalid escape sequence '\m'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/modeling/controlModel.py:175: SyntaxWarning: invalid escape sequence '\m'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/modeling/controlQoI.py:58: SyntaxWarning: invalid escape sequence '\d'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/modeling/controlQoI.py:69: SyntaxWarning: invalid escape sequence '\i'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/modeling/controlQoI.py:248: SyntaxWarning: invalid escape sequence '\O'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/modeling/meanVarRiskMeasureStochastic.py:32: SyntaxWarning: invalid escape sequence '\m'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/modeling/meanVarRiskMeasureSAA.py:38: SyntaxWarning: invalid escape sequence '\m'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/modeling/penalization.py:111: SyntaxWarning: invalid escape sequence '\s'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/modeling/penalization.py:185: SyntaxWarning: invalid escape sequence '\O'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/modeling/transformedMeanRiskMeasureSAA.py:93: SyntaxWarning: invalid escape sequence '\l'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/optimization/bfgs.py:97: SyntaxWarning: invalid escape sequence '\m'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/optimization/projectableConstraint.py:55: SyntaxWarning: invalid escape sequence '\{'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/solver/newtonBacktrackSolver.py:58: SyntaxWarning: invalid escape sequence '\i'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/solver/newtonBacktrackSolver.py:75: SyntaxWarning: invalid escape sequence '\|'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/solver/newtonBacktrackSolver.py:95: SyntaxWarning: invalid escape sequence '\i'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/solver/nonlinearVariationalSolver.py:17: SyntaxWarning: invalid escape sequence '\i'
  """
/home/jed/joss/soupy/examples/poisson/../../soupy/solver/nonlinearVariationalSolver.py:43: SyntaxWarning: invalid escape sequence '\i'
  """
```

Finally, would you like to suggest any potential reviewers? (You can mention by name or github username, but please leave out the `@` so they aren't tagged yet.)
