### 2023-11-16 f30d8e6
removed try catch from solver
