### 2023-11-16 cf314d4
updated plotting of design
