### 2023-11-16 1f5f0e8
joss-dev (#21)
* removed try catch from solver

* updated installation instructions and paper fix

* updated plotting of design
