### 2023-11-17 #5882 @dc-luo
@jedbrown Hello! Thank you for taking care of the submission and my apologies for the delayed response. The paper author affiliation has been corrected, so hopefully that builds. I've also updated the installation instructions to suggest python 3.11 for the time being.

In terms of suggestions for potential reviewers, we thought of the following
- dannys4
- dridzal
- IgorBaratta
- lheagy
- ReubenHill
