### 2023-11-30 #6101 @editorialbot
[REVIEW]: SOUPy: Stochastic PDE-constrained optimization under high-dimensional uncertainty in Python
**Submitting author:** @dc-luo (<a href="http://orcid.org/0000-0003-0466-6902">Dingcheng Luo</a>)
**Repository:** https://github.com/hippylib/soupy
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.0
