### 2024-03-07 5a757a4
changed where setLinearizationPoint is happening to avoid refactoring solves during repeated hessian evaluation
