### 2024-05-07 8ee61e0
Updated finite difference checkers to return the estimated derivatives
