### 2024-05-11 #6101 @Himscipy
Hi @tomoleary,
Thank you for submitting the paper+ software to JOSS. The developed package will be a great additions to the scientific computing research community. Following are my comments and feedback on the submission.

1.  The submitted write-up clearly states the need and highlights relevant literature.
2. I would recommend simplifying the installation of the package through containerization (docker). Currently, there are multiple dependencies for the package. Having the containerization will also ease the installation process and ensure that the library can be quickly used on HPC/large systems. For example during the installation process listed on the documentation I encountered following  error
```
conda create -n soupy -c conda-forge python=3.11 fenics==2019.1.0 petsc4py==3.19 matplotlib scipy jupyter
Retrieving notices: ...working... done
Channels:
 - conda-forge
 - defaults
Platform: osx-arm64
Solving environment: failed

PackagesNotFoundError: The following packages are not available from current channels:

  - fenics==2019.1.0

To search for alternate channels that may provide the conda package you're
looking for, navigate to

    https://anaconda.org

and use the search bar at the top of the page.

```
4. Below is how the repository compares to the community standards. If you can improve this, you could allow potential users to be major contributors to the project as well
![Screenshot 2024-05-11 at 3 54 08 PM](https://github.com/openjournals/joss-reviews/assets/7496964/993fe82e-09f2-4ae7-9964-c62d42d08ad8)

Please assist in resolving the installation issue, I have also created a issue on the repo.
