### 2024-05-14 #6101 @dc-luo
Hello @Himscipy, thank you for your initial feedback on the package!

We'll add installation instructions with docker. In the meantime, I believe the installation issue can be resolved by adding the flag `--platform osx-64` to the conda call, i.e.
`conda create -n soupy -c conda-forge python=3.11 fenics==2019.1.0 petsc4py==3.19 matplotlib scipy jupyter`
Please me know if it works for you. If so, I'll also update the current installation instruction for arm mac users.
