### 2024-05-26 31288c7
updated installation instructions for arm macos and docker
