### 2024-05-27 998f9c8
Install dev (#23)
* updated installation instructions for arm macos and docker

* updated instructions
