### 2024-05-28 #6101 @dc-luo
Hello everyone @Himscipy @IgorBaratta @jedbrown

Thank you all for your feedback on the submission! I appreciate your time in helping to improve the project.

So far, I have gone and dealt with the initial comments:

1. Installation: This was on both reviewers' checklists. I have updated the installation instructions to highlight the case of macos on arm machines. Also, based on suggestions from @Himscipy, I have also included a section on running through docker. Please see the latest [INSTALL.md](https://github.com/hippylib/soupy/blob/main/INSTALL.md) or the readthedocs for the updated installation instructions.

2. I have added an issue template and code of conduct to the repo for improvements based on community standards, as suggested by @Himscipy. Basic instructions for pull requests are available under the CONTRIBUTING.md.

I notice there are still some unchecked boxes under functionality/documentation.  Please let me know if you have any comments with respect to these or any issues you would like addressed.
