### 2024-06-21 #6101 @Himscipy
Hi @tomoleary,
Thank you for addressing the comments and revising the repository. I tried the installation fixes and it works.
I have completed my review checklist and I have no additional comments. Thank you for your contributions to the scientific computing research community with Soupy.

Best,
Himanshu
