### 2024-07-25 c88b223
JOSS: copy edits/bib
