### 2024-07-25 #6101 @jedbrown
Sorry it has taken me a long time to get to this. I have a [small PR](https://github.com/hippylib/soupy/pull/24) with minor copy edits. I think a reference for superquantiles would be nice (I'm aware of Rockafellar's work and its growing popularity in this field, but it's a term that I think will be unfamiliar to many readers of the paper). Let me know if you have any questions. If not, please proceed to
