+++ .github/workflows/run_examples.sh
@@ -5,7 +5,7 @@ export HIPPYLIB_PATH=$(pwd)/hippylib
 # Run poisson example
 cd examples/poisson
 
-python3 driver_poisson_mean.py
+python driver_poisson_mean.py
 
-mpirun -n 2 python3 driver_poisson_mean.py
+mpirun -n 2 python driver_poisson_mean.py
 
