### 2024-07-26 8a028e8
updated run scripts for using new image
