### 2024-07-26 67e1f69
adding conda
