### 2024-07-26 3701b83
Updating CI (#25)
* trying different ci image

* using python instead of python3

* updated run scripts for using new image

* adding conda
