### 2024-07-26 561b183
JOSS: copy edits/bib (#24)
