+++ paper/paper.bib
@@ -30,7 +30,7 @@
   address    = {New York, NY, USA},
   articleno  = {16},
   issue_date = {March 2021},
-  journal    = {ACM Trans. Math. Softw.},
+  journal    = {ACM Transactions on Mathematical Software},
   month      = apr,
   numpages   = {34},
   publisher  = {Association for Computing Machinery},
@@ -96,7 +96,7 @@ year = {2018}
 @Article{LiuNocedal89,
   Title                    = {On the limited memory {BFGS} methods for large scale optimization},
   Author                   = {Liu, D. C. and Nocedal, J.},
-  Journal                  = {Math. Prog.},
+  Journal                  = {Mathematical Programming},
   Year                     = {1989},
   Pages                    = {503--528},
   Volume                   = {45},
@@ -170,7 +170,7 @@ number = {2},
 issn = {0098-3500},
 url = {https://doi.org/10.1145/2566630},
 doi = {10.1145/2566630},
-journal = {ACM Trans. Math. Softw.},
+journal = {ACM Transactions on Mathematical Software},
 month = {mar},
 articleno = {9},
 numpages = {37},
