### 2024-07-26 8ac0695
Adding references to JOSS paper (#26)
* Added reference for CVaR

* update github action to build paper
