+++ paper/paper.bib
@@ -217,3 +217,15 @@ url = {https://trilinos.github.io/rol.html}
      month = {01},
      url = {https://trilinos.github.io/pdfs/ROL.pdf}
 }
+
+@Article{RockafellarUryasev00,
+  author    = {R. Tyrrell Rockafellar and Stanislav Uryasev},
+  journal   = {The Journal of Risk},
+  title     = {Optimization of conditional value-at-risk},
+  year      = {2000},
+  number    = {3},
+  pages     = {21--41},
+  volume    = {2},
+  doi       = {10.21314/jor.2000.038},
+  publisher = {Infopro Digital Services Limited},
+}
