### 2024-07-26 #6101 @dc-luo
Thanks @jedbrown! I've merged the PR and added a reference to superquantile. Would you like to check through the updated paper before I archive the repo?
