### 2024-07-26 f2627d1
Update CHANGELOG.md
