+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-Development version
+Version 1.1.0, released on Jul 26, 2024
 ----------------------------------------
 - Fixed bug for gradient norm calculation inside `ControlModel`
 - Added support for generic risk measures and chance constrained optimization through `TransformedMeanRiskMeasureSAA` and `PenalizationControlcostFunctional`
-- Added jupyter notebook tutorials 
+- Added jupyter notebook tutorials
+- Updates to installation instructions
+- JOSS paper submission
 
 Version 1.0.0, released on Aug 11, 2023
 ----------------------------------------
