### 2024-07-26 #6101 @dc-luo
Thanks @jedbrown!

I've gone and versioned the reviewed software under the tag `v1.1.0`. Here's the doi for the archived version on zenodo [10.5281/zenodo.12997883](https://doi.org/10.5281/zenodo.12997883), which now has matching metadata info to the JOSS paper.

Please let me know if there's anything else to do.
