### 2024-07-26 #6101 @danielskatz
@dc-luo - I'm suggesting some changes in https://github.com/hippylib/soupy/pull/27. They are all straightforward, except perhaps the one for the website in the bib, where I'm suggesting using an archival version of the website that will always be preserved rather than an actual website with a reference date. The version that I link to is the same as it was in Sept 2023, as far as I can tell.
