### 2024-07-28 0233d1b
changes in JOSS paper (#27)
* minor changes in paper.md

* small changes in paper.bib
