+++ paper/paper.bib
@@ -22,7 +22,7 @@
 
 @article{VillaPetraGhattas21,
   author     = {Villa, Umberto and Petra, Noemi and Ghattas, Omar},
-  title      = {{HIPPYlib: An Extensible Software Framework for Large-Scale Inverse Problems Governed by PDEs: Part I: Deterministic Inversion and Linearized Bayesian Inference}},
+  title      = {{HIPPYlib}: An Extensible Software Framework for Large-Scale Inverse Problems Governed by {PDE}s: {P}art {I}: Deterministic Inversion and Linearized {B}ayesian Inference},
   doi        = {10.1145/3428447},
   issn       = {0098-3500},
   number     = {2},
@@ -82,8 +82,8 @@ year = {2018}
             Harris, Charles R. and Archibald, Anne M. and
             Ribeiro, Ant{\^o}nio H. and Pedregosa, Fabian and
             {van Mulbregt}, Paul and {SciPy 1.0 Contributors}},
-  title   = {{{SciPy} 1.0: Fundamental Algorithms for Scientific
-            Computing in Python}},
+  title   = {{SciPy} 1.0: Fundamental Algorithms for Scientific
+            Computing in {P}ython},
   journal = {Nature Methods},
   year    = {2020},
   volume  = {17},
@@ -178,10 +178,10 @@ numpages = {37},
 
 
 @Manual{trilinos-website,
-title = {The {T}rilinos {P}roject {W}ebsite},
-author = {{The Trilinos Project Team}},
-year = {2023 (acccessed September 13, 2023)},
-url = {https://trilinos.github.io/rol.html}
+title = {{T}rilinos {P}roject Website},
+author = {{Trilinos Project Team}},
+year = {2024},
+url = {https://web.archive.org/web/20240228185301/https://trilinos.github.io/rol.html}
 }
 
 
