### 2024-07-28 #6101 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
